refactor(client): some components have been merged and useless conditions have been deleted
Some imports and codes that were not being used were removed and the revisionquest component was
merged with revisionresponse and the addquest with addresponse generating the Revision and add
files.
