+++ client/src/pages/AddCard.tsx
@@ -1,18 +1,15 @@
 import { getCard } from '../service/api'
 import { useEffect, useState} from 'react'
-import React from 'react'
 import { ICardOrdered } from '../domain/useCase/orderCard'
 import { OrderCardService } from '../service/OrderCardService'
-import AddQuest  from '../components/AddQuest'
-import AddResponse  from '../components/AddResponse'
-import { Link, Route, Router, Switch, useRouteMatch } from 'react-router-dom'
-const CardContext = React.createContext('');
-
+import AddQuest  from '../components/Add'
+import AddResponse  from '../components/Add'
+import { Link, Route, Switch, useRouteMatch } from 'react-router-dom'
 
 const AddCard = () => { 
-    let match = useRouteMatch();
     const [orderCards, setOrderCards ] = useState<ICardOrdered[]>([])
     const [index, setIndex] = useState(0)
+    let match = useRouteMatch();
     
     useEffect(() => {
         async function downloadCards() {
@@ -41,7 +38,6 @@ const AddCard = () => {
                         </Route>
                         <Route path={`${match.path}/Response`}>
                             <AddResponse/>
-                            {console.log(match.path)}
                         </Route>
                     </Switch>
             </main>
