+++ client/src/pages/Dashboard.tsx
@@ -1,20 +1,15 @@
-import RevisionQuest from '../components/RevisionQuest'
-import RevisionResponse from '../components/RevisionResponse'
+import RevisionQuest from '../components/Revision'
+import RevisionResponse from '../components/Revision'
 import { getCard } from '../service/api'
 import { useEffect, useState} from 'react'
-import React from 'react'
 import { ICardOrdered } from '../domain/useCase/orderCard'
 import { OrderCardService } from '../service/OrderCardService'
-import AddCard  from '../components/AddQuest'
 import { Route, Switch, useRouteMatch } from 'react-router-dom'
-const CardContext = React.createContext('');
-
-// let b = new history('test2');
 
 const Dashboard = () => { 
-    let match = useRouteMatch();
     const [orderCards, setOrderCards ] = useState<ICardOrdered[]>([])
     const [index, setIndex] = useState(0)
+    let match = useRouteMatch();
     
     useEffect(() => {
         async function downloadCards() {
@@ -27,12 +22,6 @@ const Dashboard = () => {
         downloadCards()
 
     }, [])
-    function resetCard() {
-        setIndex(index + 1)
-        document.getElementById("dashboard__content__back")!.style.display = "none"
-        document.getElementById("dashboard__content__seeBack")!.style.display = "initial"
-        
-    }
 
 	return(
         <div className="container" id="containerCard">
